updates adding items to cart
